Added tests for add and delete functions
